Changed IMDBSearch.py to imdb_search.py to conform with PEP8.
Edited functions within imdb_search.py module such that they would have the optional "override" string, which would skip the process of asking for user input.

Began implementing test case files. The first case searches "The Aviator" on override. "The Aviator" has a summary which contains an <a> tag, which currently breeches the element hierarchy and causes the summary to cut off at the beginning of the <a> tag, despite having a longer summary. Further experimentation with this test case will determine the process of circumventing this edge case.