Implementation of Collection class
First commit of the Collection class in collection.py. This should allow
the creation of collections of items. So far, the only compatible items
are the Movie objects in imdb_search.py.

Many more changes to come. Implementation is very, very rough, and there
could be a ton missing from the class.